Add camera_trace and fetch_request
