+++ src/vvplay_async_prefetch/mod.rs
@@ -1,1 +1,3 @@
 pub mod network_trace;
+pub mod camera_trace;
+pub mod fetch_request;
